+++ src/gw/auth.py
@@ -32,6 +32,7 @@ DEFAULT_SCOPES = [
     "https://www.googleapis.com/auth/gmail.modify",
     "https://www.googleapis.com/auth/gmail.send",
     "https://www.googleapis.com/auth/drive",
+    "https://www.googleapis.com/auth/tasks",
     "https://www.googleapis.com/auth/spreadsheets",
     "https://www.googleapis.com/auth/documents.readonly",
     "https://www.googleapis.com/auth/contacts.readonly",